guides/diffusers/examples/requirements.txt
# Hugging Face Diffusers with Flux Models - Requirements
# Updated for FLUX.1-Kontext-dev support

# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.30.0
transformers>=4.37.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
huggingface-hub>=0.20.0

# Optional: for enhanced performance
# xformers>=0.0.22  # Uncomment for memory optimization
# bitsandbytes>=0.41.0  # Uncomment for 8-bit optimization

# CLI utilities
click>=8.1.0
rich>=13.0.0